post a new article
This project is inspired by the Timescape exhibit in the 9/11 Memorial
Museum. I used the tools and techniques of data science to investigate
news topics that emerged from the events of September 11 and track their
newsworthiness in the years since. I used the results of my analysis to
build an interactive visualization so anyone can explore these topics
and find related articles.